git-svn-id: svn://202.107.200.162:8002/repo/mcs@6344 e790cb50-a368-4814-8953-9b4d82950c0f

+++ src/com/lierda/web/service/ZBuildingServiceI.java
@@ -19,7 +19,7 @@ public interface ZBuildingServiceI extends CommonService{
 	
 	public Map<String, Object> getTotalPower(Map<String, Object> currentPower);
 	
-	public Map<String, double[]> getPowerByType(List<PowerRecordingEntity> recordingEntities,Map<String, Object> currentPower);
+	public Map<String, float[]> getPowerByType(List<PowerRecordingEntity> recordingEntities,Map<String, Object> currentPower);
 	
 	public Map<String, Object> getCurrentPower(long timeStart,Map<String, String> powerMap);
